:art: Rename: 서비스 계층 내에서 command, query 패키지 분할
