+++ src/main/java/site/moamoa/backend/service/memberpost/command/MemberPostCommandServiceImpl.java
@@ -1,9 +1,10 @@
-package site.moamoa.backend.service.memberpost;
+package site.moamoa.backend.service.memberpost.command;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import site.moamoa.backend.domain.mapping.MemberPost;
 import site.moamoa.backend.repository.memberpost.MemberPostRepository;
+import site.moamoa.backend.service.memberpost.command.MemberPostCommandService;
 
 @AllArgsConstructor
 @Service
