+++ src/main/java/site/moamoa/backend/service/memberpost/query/MemberPostQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.service.memberpost;
+package site.moamoa.backend.service.memberpost.query;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -7,6 +7,7 @@ import site.moamoa.backend.domain.mapping.MemberPost;
 import site.moamoa.backend.exception.member.MemberIsNotAuthorException;
 import site.moamoa.backend.exception.memberpost.MemberPostNotFoundException;
 import site.moamoa.backend.repository.memberpost.MemberPostRepository;
+import site.moamoa.backend.service.memberpost.query.MemberPostQueryService;
 
 @Service
 @AllArgsConstructor
